Modified

+++ src/Resource.java
@@ -15,7 +15,7 @@ public abstract class Resource implements Serializable{
 	protected LocalTime startTime,endTime;
 	protected String location;
 	protected userType resourceUserType;
-	protected long Allowance_time;
+	protected int Allowance_time;
 	public Resource() {
 		
 	}
@@ -95,10 +95,10 @@ public abstract class Resource implements Serializable{
 	public String toString(){
 		return ("id: "+this.ID + " Location: "+ this.location + " Start Date: " + this.startDate +" End Date: "+this.endDate+" Start Time: "+this.startTime + " End Time: "+ this.endTime +" User Type: " + this.resourceUserType +" Allowance time="+ this.Allowance_time);
   	}
-	public long getAllowance_time() {
+	public int getAllowance_time() {
 		return Allowance_time;
 	}
-	public void setAllowance_time(long allowance_time) {
+	public void setAllowance_time(int allowance_time) {
 		Allowance_time = allowance_time;
 	}
